Moved chat classes into chats directory
